Refactor in CodeConverters to use base class to handle Accessibility
