+++ FigmaSharp.NativeControls/FigmaSharp.NativeControls.Cocoa/Converters/ComboBoxConverter.cs
@@ -82,7 +82,7 @@ namespace FigmaSharp.NativeControls.Cocoa
 			return new View (view);
 		}
 
-		public override string ConvertToCode (FigmaCodeNode currentNode, FigmaCodeNode parentNode, FigmaCodeRendererService rendererService)
+		protected override StringBuilder OnConvertToCode(FigmaCodeNode currentNode, FigmaCodeNode parentNode, FigmaCodeRendererService rendererService)
 		{
 			var figmaInstance = (FigmaFrameEntity)currentNode.Node;
 
@@ -118,7 +118,7 @@ namespace FigmaSharp.NativeControls.Cocoa
 			//	builder.AppendLine (string.Format ("{0}.Appearance = NSAppearance.GetAppearance ({1});", name, NSAppearance.NameDarkAqua.GetType ().FullName));
 			//}
 
-			return builder.ToString ();
+			return builder;
 		}
 	}
 }
