+++ FigmaSharp.NativeControls/FigmaSharp.NativeControls.Cocoa/Converters/SpinnerConverter.cs
@@ -73,7 +73,7 @@ namespace FigmaSharp.NativeControls.Cocoa
 			return view;
 		}
 
-		public override string ConvertToCode (FigmaCodeNode currentNode, FigmaCodeNode parentNode, FigmaCodeRendererService rendererService)
+		protected override StringBuilder OnConvertToCode (FigmaCodeNode currentNode, FigmaCodeNode parentNode, FigmaCodeRendererService rendererService)
 		{
 			var figmaInstance = (FigmaFrameEntity)currentNode.Node;
 
@@ -103,7 +103,7 @@ namespace FigmaSharp.NativeControls.Cocoa
 					break;
 			}
 
-			return builder.ToString ();
+			return builder;
 		}
 	}
 }
