+++ FigmaSharp.NativeControls/FigmaSharp.NativeControls.Cocoa/Converters/StepperConverter.cs
@@ -75,7 +75,7 @@ namespace FigmaSharp.NativeControls.Cocoa
 			return view;
 		}
 
-		public override string ConvertToCode (FigmaCodeNode currentNode, FigmaCodeNode parentNode, FigmaCodeRendererService rendererService)
+		protected override StringBuilder OnConvertToCode(FigmaCodeNode currentNode, FigmaCodeNode parentNode, FigmaCodeRendererService rendererService)
 		{
 			var figmaInstance = (FigmaFrameEntity)currentNode.Node;
 
@@ -99,7 +99,7 @@ namespace FigmaSharp.NativeControls.Cocoa
 					break;
 			}
 
-			return builder.ToString ();
+			return builder;
 		}
 
 	}
